Refactored the model REST API for POST and PUT.
A Model contains both a definition and a list of data items.
When POSTing to /models you need to provide a definition and some data items, which changes the API contract we have when we PUT on /models/:model-id/definition.

You also can pass some data there, which will be created if they validate the definition.

r=@Natim